ai/backend/requirements.txt
fastapi==0.119.0
uvicorn==0.38.0
python-dotenv==1.1.1
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.185.0
pydantic==2.12.3
langchain-google-genai==2.1.12
langchain-core==0.3.79
protobuf==4.25.1
